requirements.txt
google-genai==0.5.0
llama-index==0.12.11
llama-index-llms-gemini==0.4.3
llama-index-embeddings-gemini==0.3.1
websockets
python-dotenv
google-cloud-vision
pymupdf==1.24.10
supabase
